+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ModInterface.lua
@@ -56,7 +56,7 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
             --  LOAD PRE-EXISTING CONFIGURATION
             --  -------------------------------
 
-            for modName, content in pairs(Ext.JsonParse(Ext.LoadFile("S7_ConfigData.json") or "{}")) do
+            for modName, content in pairs(Ext.JsonParse(Ext.LoadFile("S7_ConfigData.json") or {})) do
                 if modName == quickMenuVars.modName then --  if configuration already exists.
                     quickMenuVars.configData = Ext.JsonParse(content) --  Pre-load configData with existing configuration.
                 end